Refacto : change mattermost lib package hierarchie
