+++ src/schedulers/mattermostScheduler/index.ts
@@ -62,3 +62,5 @@ export const getMattermostUsersActiveGithubUsersNotInTeam = async (teamId: strin
   export * from './moveUsersToAlumniTeam'
   export * from './reactivateUsers'
   export * from './removeUsersFromCommunityTeam'
+  export * from './addUsersNotInCommunityToCommunityTeam'
+  export * from './syncMattermostUserWithMattermostMemberInfosTable'
